Moved typescript declaration file to root folder and deleted unnecessary tsconfig.json file. Restructured declaration file to make the type detectable.
